+++ search/network/openmp/auxi.h
@@ -36,7 +36,7 @@ void triginterp (complex double *ya, complex double *yb, double *shftf,
 		 int N, int nfft, complex double *outa, complex double *outb);
 double var (double *, int);
 
-void gridr (double *, int *, int *, int *, double, double);
+void gridr (double *, float *, float *, float *, double, double);
 double FStat (FLOAT_TYPE *, int, int, int);
 
 int ludcmp (double *, int, int *, double *);
